cli: chown before reporting
User must have write access to the reported directory to be able to
report it but abrt-dbus allows the user to read data of problems that
belongs to him which may not be accessible in file system.

The GUI does the same and make sures the user can write to the reported
directory by chowning it before reporting.

Signed-off-by: Jakub Filak <jfilak@redhat.com>