a-a-p-ccpp-analysis: use ask_yes_no_save_result instead of ask_yes_no_yesforever
We change this function because we want users have an opportunity to choose
option never upload the coredump and always generate the backtrace locally.

We do not make the migration to this new feature because we believe the user will
be ok with one more click to configure this again.
We also thing that the re-configuraton will help will help familiarize users
with the new feature.

Related to rhbz#986876

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>