spec: remove PyGObject from all Requires
pygobject3 was in the Requires of a Python 3 package, but pygobject3 is
a Python 2 module that provides a wrapper for the GObject library
(Gtk+3). So, this Require can be removed as Python 3 do not use Python 2
modules.

pygobject2 was added in commit ee7619c1c729b63c4ac0b21dc88692cbf042c5a3
without any explanation and I do think we do not need it, because abrt
does not use PyGTK (Gtk+2) and the Python 3 sub-packages live without
PyGObject for a quite long time (see the first paragraph).

Signed-off-by: Jakub Filak <jfilak@redhat.com>