cli: use internal command impl in the command process
It did not seem to be a good idea to add wrappers for the internal
commands, because the wrappers would be one line functions. Now, we need
to do more sophisticated processing (authenticate, chown), so adding the
wrappers is the best choice to provide the same functionality in the
command process.

Signed-off-by: Jakub Filak <jfilak@redhat.com>