testsuite: make oops-processing distro specific
There is no "DropNotReportableOopses" option in oops.conf on rhel, so remove
the test cases related to this.

Remove the "oops_not_reportable_no_reliable_frame.test" test case because if
there is not "DropNotReportableOopses = no" option the dump dir is deleted
bacause of bad backtrace and the test fails.

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>