lib: introduce the SysRootPath configuration option
We need a way how to tell all ABRT tools that the root file system of
the current system is not their root file system.

This option comes in handy when running ABRT software in a container
where ABRT must gather system details from the host and not from the
container.

In the history, I was adding the '--chroot' argument to the tools that
were in need to be aware of the correct root file system, but this
approach brings the need to configure a lot of stuff. So the new option
can be considered a convenient way how to tell all ABRT tools that the
root file system is somewhere else.

Signed-off-by: Jakub Filak <jakub@thefilaks.net>