packages: honor SysRootPath in tools saving pkg details
Start using the new SysRootPath option that tells ABRT tools where the
root file system lives.

It is already possible to run these tools with the -r argument that
fulfills the same role as the SysRootPath configuration option. However,
keep the track of tools that need to be invoked with a special argument is
not very convenient and is error prone. The best way is to properly
configured the SysRootPath configuration option and let all ABRT tools
to use it.

Signed-off-by: Jakub Filak <jakub@thefilaks.net>