log watchers: honor SysRootPath
The kernel oops addon and the XOrg addon doe not gather problem data
from a process, thus these addons cannot find out that the problem occurred
in a changed root environment. One may that this is not a problem as the
logged problems occur in the default namespace, but we must no forget
that the addon itself might be used from a non-default namespace.

The SysRooPath configuration option has been introduced to enable the
addons to gather correct problem data if started in a container. This
commit just enhances the log watchers with saving of SysRootPath in the
problem data.

Signed-off-by: Jakub Filak <jakub@thefilaks.net>