a-a-s-c-data: container's root path as seen from host
This commit adds abrt-action-save-container-data ability to find and save
the file system path of a container's root directory as seen from a
host. The path will be saved in a file named 'container_sysroot'.

Currently only Docker is supported.

This task requires two steps. First, we need to find the source device
of the '/' mount point in the crashed process' /proc/[pid]/mount_info
file. Then we need to check if the source device is mounted somewhere in
the host. We also need to check if a mount point for the source device
uses the same root directory inside the source device as the crashed
process container.

Signed-off-by: Jakub Filak <jakub@thefilaks.net>